--- main_esp32_meter/README.md
@@ -1,11 +1,0 @@
-# SmartEnergyMeter_RGB
-Smart Energy Meter with SIM900 and Current Sensor
-
-  Project "Smart Energy Meter" using current sensor arduino and GSM Module and a display
-  To monitor energy consumption and cost and delelop a pre-paid system with notification and digital control of sort...
-  Developed by Sakib Ahmed Sumdany, Adnan Sabbir
-  November 14, 2016.
-  Released into the public domain.
-  Dependencies: ACS712 Library by Sakib Ahmed
-  
- https://www.youtube.com/watch?v=GyTl6QiHSSM
